+++ src/Entities/WebhookCredential.php
@@ -19,19 +19,13 @@ use Rebilly\Rest\Entity;
 final class WebhookCredential extends Entity
 {
     /**
-     * @return string
-     */
-    public function getId()
-    {
-        return $this->getHash();
-    }
-
-    /**
+     * @deprecated use {@see getId()} instead
+     *
      * @return string
      */
     public function getHash()
     {
-        return $this->getAttribute('hash');
+        return $this->getId();
     }
 
     /**